refactor the code and added component folder.
